# docs/swagger.yaml
basePath: /
definitions:
  types.ABI:
    properties:
      abi_content:
        description: ABI JSON内容
        type: string
      created_at:
        type: string
      description:
        description: ABI描述
        type: string
      id:
        type: integer
      is_shared:
        description: 是否为共享ABI
        type: boolean
      name:
        description: ABI名称
        type: string
      owner:
        description: 所有者地址，全0表示共享ABI
        type: string
      updated_at:
        type: string
    type: object
  types.ABIListResponse:
    properties:
      shared_abis:
        description: 平台共享的ABI
        items:
          $ref: '#/definitions/types.ABI'
        type: array
      user_abis:
        description: 用户创建的ABI
        items:
          $ref: '#/definitions/types.ABI'
        type: array
    type: object
  types.ABIResponse:
    properties:
      abi_content:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_shared:
        type: boolean
      name:
        type: string
      owner:
        type: string
      updated_at:
        type: string
    type: object
  types.ABIValidationResult:
    properties:
      error_message:
        type: string
      event_count:
        type: integer
      function_count:
        type: integer
      is_valid:
        type: boolean
      warnings:
        items:
          type: string
        type: array
    type: object
  types.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  types.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/types.APIError'
      success:
        type: boolean
    type: object
...
  types.CreateABIRequest:
    properties:
      abi_content:
        type: string
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - abi_content
    - name
    type: object
...
  types.DeleteABIRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  types.GetABIByIDRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  types.UpdateABIWithIDRequest:
    properties:
      abi_content:
        type: string
      description:
        maxLength: 500
        type: string
      id:
        type: integer
      name:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - abi_content
    - id
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: TimeLocker Backend API
  title: TimeLocker Backend API
  version: "1.0"
paths:
  /api/v1/abi:
    post:
      consumes:
      - application/json
      description: 用户创建新的智能合约ABI。系统会验证ABI格式的正确性。每个用户在同一名称下只能创建一个ABI。名称长度1-200；描述≤500。
      parameters:
      - description: 创建ABI请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.CreateABIRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ABI创建成功
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.ABIResponse'
              type: object
        "400":
          description: 请求参数错误或ABI格式无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "409":
          description: ABI名称已存在
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "422":
          description: 参数校验失败
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 创建ABI
      tags:
      - ABI
  /api/v1/abi/delete:
    post:
      consumes:
      - application/json
      description: 删除用户创建的ABI。用户只能删除自己创建的ABI，不能删除平台共享的ABI。删除操作是不可逆的。
      parameters:
      - description: 删除ABI请求体（包含ID）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.DeleteABIRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ABI删除成功
          schema:
            $ref: '#/definitions/types.APIResponse'
        "400":
          description: 无效的ABI ID
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "403":
          description: 无权删除该ABI或尝试删除共享ABI
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "404":
          description: ABI不存在
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 删除ABI
      tags:
      - ABI
  /api/v1/abi/get:
    post:
      consumes:
      - application/json
      description: 根据ABI ID获取详细信息。用户只能访问自己创建的ABI或平台共享的ABI。
      parameters:
      - description: 获取ABI详情请求体（包含ID）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.GetABIByIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 获取ABI详情成功
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.ABIResponse'
              type: object
        "400":
          description: 无效的ABI ID
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "403":
          description: 无权访问该ABI
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "404":
          description: ABI不存在
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 获取ABI详情
      tags:
      - ABI
  /api/v1/abi/list:
    post:
      consumes:
      - application/json
      description: 获取用户可访问的ABI列表，包括用户自己创建的ABI和平台共享的ABI。返回结果分为两个部分：用户ABI和共享ABI。
      produces:
      - application/json
      responses:
        "200":
          description: 获取ABI列表成功
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.ABIListResponse'
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 获取ABI列表
      tags:
      - ABI
  /api/v1/abi/update:
    post:
      consumes:
      - application/json
      description: 更新用户创建的ABI。系统会重新验证ABI格式。用户只能更新自己创建的ABI。名称长度1-200；描述≤500。
      parameters:
      - description: 更新ABI请求体（包含ID）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.UpdateABIWithIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ABI更新成功
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.ABIResponse'
              type: object
        "400":
          description: 请求参数错误或ABI格式无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "403":
          description: 无权更新该ABI
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "404":
          description: ABI不存在
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "409":
          description: ABI名称已存在
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "422":
          description: 参数校验失败
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 更新ABI
      tags:
      - ABI
  /api/v1/abi/validate:
    post:
      consumes:
      - application/json
      description: 验证智能合约ABI的格式正确性，返回详细的验证结果。此接口可用于在创建或更新ABI前进行预验证。
      parameters:
      - description: 验证ABI请求体
        in: body
        name: request
        required: true
        schema:
          properties:
            abi_content:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: ABI验证完成
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.ABIValidationResult'
              type: object
        "400":
          description: 请求参数错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "422":
          description: 参数校验失败
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 验证ABI格式
      tags:
      - ABI
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"